Moving DbRegistration
