+++ Tests/BackendTest/Mongo2GoFixture.cs
@@ -1,6 +1,5 @@
 using System;
-using System.Collections.Generic;
-using System.Text;
+using Backend;
 using Mongo2Go;
 using MongoDB.Driver;
 using Xunit;
